Add `plugins.md` which holds history and possible new ones
I tend to forget what I have tried over the years and why I stopped
using them. I started documenting it on the included configuration, but
now I want a pretty clean configuration which just the used stuff, and
the history and the reasons will go in this new file as well as the
plugins I catch around and wanna try but don't have time yet.